backend/pyproject.toml
[tool.poetry]
name = "ipr"
version = "0.1.0"
description = ""
authors = ["Puria Nafisi Azizi <puria@dyne.org>"]

[tool.poetry.scripts]
start = "ipr.app:start"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.75.2"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
reuse = "^0.14.0"
scancode-toolkit = "^30.1.0"
sh = "^1.14.2"
SQLAlchemy = "^1.4.35"
requests = "^2.27.1"
GitPython = "^3.1.27"
celery = "^5.2.6"
redis = "^4.2.2"
pastel = "^0.2.1"
psycopg2-binary = "^2.9.3"
alembic = "^1.7.7"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = {version = "^22.3.0", allow-prereleases = true}
pytest-cov = "^3.0.0"
flower = "^1.0.0"
locust = "^2.8.6"
pastel = "^0.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
